almost ch7

+++ controllers/image.js
@@ -1,39 +1,38 @@
 const fs = require('fs'),
   path = require('path'),
-  sidebar = require('../helpers/sidebar');
-
+  sidebar = require('../helpers/sidebar'),
+  Models = require('../models'),
+  md5 = require('md5');
 
 module.exports = {
   index(req, res) {
-    const ViewModel = {
-      image: {
-        uniqueId: 1,
-        title: 'Sample Image 1',
-        description: 'This is a sample.',
-        filename: 'sample1.jpg',
-        Views: 0,
-        likes: 0,
-        timestamp: Date.now()
-      },
-      comments: [{
-        image_id: 1,
-        email: 'test@testing.com',
-        name: 'Test Tester',
-        gravatar: 'http://lorempixel.com/75/75/animals/1',
-        comment: 'This is a test comment...',
-        timestamp: Date.now()
-      }, {
-        image_id: 1,
-        email: 'test@testing.com',
-        name: 'Test Tester',
-        gravatar: 'http://lorempixel.com/75/75/animals/2',
-        comment: 'Another followup comment!',
-        timestamp: Date.now()
-      }]
+    var viewModel = {
+      image: {},
+      comments: []
     };
-    sidebar(ViewModel, (ViewModel) => {
-      res.render('image', ViewModel);
-    })
+
+    Models.Image.findOne({ filename: { $regex: req.params.image_id } },
+      (err, image) => {
+        if (err) { throw err; }
+        if (image) {
+          image.views = image.views + 1;
+          viewModel.image = image;
+          image.save();
+
+          Models.Comment.find({ image_id: image._id }, {},
+            { sort: { 'timestamp': 1 } },
+            (err, comments) => {
+              if (err) { throw err; }
+              viewModel.comments = comments;
+              sidebar(viewModel, (viewModel) => {
+                res.render('image', viewModel);
+              });
+            }
+          );
+        } else {
+          res.redirect('/');
+        }
+      });
     ;
   },
   create(req, res) {
@@ -45,28 +44,72 @@ module.exports = {
         imgUrl += possible.charAt(Math.floor(Math.random() * possible.length));
       }
 
-      const tempPath = req.file.path,
-        ext = path.extname(req.file.originalname).toLowerCase(),
-        targetPath = path.resolve(`./public/upload/${ imgUrl }${ ext }`);
+      Models.Image.find({ filename: imgUrl }, (err, images) => {
+        if (images.length > 0) {
+          // if a matching image was found, try again (start over):
+          saveImage();
+        } else {
+          var tempPath = req.file.path,
+            ext = path.extname(req.file.originalname).toLowerCase(),
+            targetPath = path.resolve(`./public/upload/${ imgUrl }${ ext }`);
 
-      if (ext === '.png' || ext === '.jpg' || ext === '.jpeg' || ext === '.gif') {
-        fs.rename(tempPath, targetPath, (err) => {
-          if (err) throw err;
-          res.redirect(`/images/${imgUrl}`);
-        });
-      } else {
-        fs.unlink(tempPath, (err) => {
-          if (err) throw err;
-          res.json(500, { error: 'Only image files are allowed.' });
-        });
-      }
+          if (ext === '.png' || ext === '.jpg' || ext === '.jpeg' || ext === '.gif') {
+            fs.rename(tempPath, targetPath, (err) => {
+              if (err) throw err;
+              var newImg = new Models.Image({
+                title: req.body.title,
+                filename: imgUrl + ext,
+                description: req.body.description
+              });
+              // and save the new Image
+              newImg.save((err, image) => {
+                res.redirect(`/images/${image.uniqueId}`);
+              });
+            });
+          } else {
+            fs.unlink(tempPath, (err) => {
+              if (err) throw err;
+              res.json(500, { error: 'Only image files are allowed.' });
+            });
+          }
+        }
+      });
     };
-    saveImage();
+    saveImage()
   },
   like(req, res) {
-    res.json({ likes: 1 });
+    Models.Image.findOne({
+        filename: { $regex: req.params.image_id }
+      },
+      (err, image) => {
+        if (!err && image) {
+          image.likes = image.likes + 1;
+          image.save((err) => {
+            if (err) {
+              res.json(err);
+            } else {
+              res.json({ likes: image.likes });
+            }
+          });
+        }
+      });
   },
   comment(req, res) {
-    res.send('The image:comment POST controller');
+    Models.Image.findOne({
+        filename: { $regex: req.params.image_id }
+      },
+      (err, image) => {
+        if (!err && image) {
+          var newComment = new Models.Comment(req.body);
+          newComment.gravatar = md5(newComment.email);
+          newComment.image_id = image._id;
+          newComment.save((err, comment) => {
+            if (err) { throw err; }
+            res.redirect(`/images/${image.uniqueId}#${comment._id}`);
+          });
+        } else {
+          res.redirect('/');
+        }
+      });
   }
 };
